sui_sdk_types

Struct Bitmap

pub struct Bitmap(/* private fields */);

A compressed bitmap using the Roaring bitmap compression scheme.

BCS

The BCS serialized form for this type is defined by the following ABNF:

roaring-bitmap = bytes  ; where the contents of the bytes are valid
                        ; according to the serialized spec for
                        ; roaring bitmaps

Implementations

impl Bitmap

pub fn new() -> Self

Creates an empty Bitmap.

Examples

use sui_sdk_types::Bitmap;
let bitmap = Bitmap::new();

pub fn len(&self) -> u64

Returns the number of distinct integers added to the set.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
assert_eq!(bitmap.len(), 0);

bitmap.insert(3);
assert_eq!(bitmap.len(), 1);

bitmap.insert(3);
bitmap.insert(4);
assert_eq!(bitmap.len(), 2);

pub fn clear(&mut self)

Clears all integers in this set.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
bitmap.insert(1);
assert_eq!(bitmap.contains(1), true);
bitmap.clear();
assert_eq!(bitmap.contains(1), false);

pub fn is_empty(&self) -> bool

Returns true if there are no integers in this set.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
assert_eq!(bitmap.is_empty(), true);

bitmap.insert(3);
assert_eq!(bitmap.is_empty(), false);

pub fn insert(&mut self, value: u32) -> bool

Adds a value to the set.

Returns whether the value was absent from the set.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
assert_eq!(bitmap.insert(3), true);
assert_eq!(bitmap.insert(3), false);
assert_eq!(bitmap.contains(3), true);

pub fn insert_range<R>(&mut self, range: R) -> u64
where R: RangeBounds<u32>,

Inserts a range of values. Returns the number of inserted values.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
bitmap.insert_range(2..4);
assert!(bitmap.contains(2));
assert!(bitmap.contains(3));
assert!(!bitmap.contains(4));

pub fn remove(&mut self, value: u32) -> bool

Removes a value from the set. Returns true if the value was present in the set.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
bitmap.insert(3);
assert_eq!(bitmap.remove(3), true);
assert_eq!(bitmap.remove(3), false);
assert_eq!(bitmap.contains(3), false);

pub fn contains(&self, value: u32) -> bool

Returns true if this set contains the specified integer.

Examples

use sui_sdk_types::Bitmap;

let mut bitmap = Bitmap::new();
bitmap.insert(1);
assert_eq!(bitmap.contains(0), false);
assert_eq!(bitmap.contains(1), true);
assert_eq!(bitmap.contains(100), false);

pub fn iter(&self) -> impl Iterator<Item = u32> + use<'_>

Iterator over each value stored in the Bitmap, guarantees values are ordered by value.

Examples

use core::iter::FromIterator;
use sui_sdk_types::Bitmap;

let bitmap = (1..3).collect::<Bitmap>();
let mut iter = bitmap.iter();

assert_eq!(iter.next(), Some(1));
assert_eq!(iter.next(), Some(2));
assert_eq!(iter.next(), None);

pub fn deserialize_from<R: Read>(reader: R) -> Result<Self>

Available on crate feature serde only.

Deserialize a bitmap into memory from the standard Roaring on-disk format. This is compatible with the official C/C++, Java and Go implementations. This method checks that all of the internal values are valid.

Examples

use sui_sdk_types::Bitmap;

let bitmap1: Bitmap = (1..4).collect();
let mut bytes = vec![];
bitmap1.serialize_into(&mut bytes).unwrap();
let bitmap2 = Bitmap::deserialize_from(&bytes[..]).unwrap();

assert_eq!(bitmap1, bitmap2);

pub fn serialize_into<W: Write>(&self, writer: W) -> Result<()>

Available on crate feature serde only.

Serialize this bitmap into the standard Roaring on-disk format. This is compatible with the official C/C++, Java and Go implementations.

Examples

use sui_sdk_types::Bitmap;

let bitmap1: Bitmap = (1..4).collect();
let mut bytes = vec![];
bitmap1.serialize_into(&mut bytes).unwrap();
let bitmap2 = Bitmap::deserialize_from(&bytes[..]).unwrap();

assert_eq!(bitmap1, bitmap2);