Module sui_sdk_types::types

Modules

framework

Rust definitions of move/sui framework types.

Structs

ActiveJwk

Address

AuthenticatorStateExpire

AuthenticatorStateUpdate

BalanceChange

Bls12381PublicKey

Bls12381Signature

Bn254FieldElement

CancelledTransaction

ChangeEpoch

ChangedObject

CheckpointContents

CheckpointContentsDigest

CheckpointData

CheckpointDigest

CheckpointSummary

CheckpointTransaction

CheckpointTransactionInfo

CircomG1

A G1 point in BN254 serialized as a vector of three strings which is the canonical decimal representation of the projective coordinates in Fq.

CircomG2

A G2 point in BN254 serialized as a vector of three vectors each being a vector of two strings which are the canonical decimal representation of the coefficients of the projective coordinates in Fq2.

Claim

A claim consists of value and index_mod_4.

ConsensusCommitDigest

ConsensusCommitPrologue

Only commit_timestamp_ms is passed to the move call currently. However we include epoch and round to make sure each ConsensusCommitPrologue has a unique tx digest.

ConsensusCommitPrologueV2

ConsensusCommitPrologueV3

Digest

A representation of a 32 byte digest

DigestParseError

Ed25519PublicKey

Ed25519Signature

EffectsAuxiliaryDataDigest

EffectsObjectChange

EndOfEpochData

Event

Specific type of event

GasCostSummary

Summary of gas charges.

GasPayment

GenesisObject

GenesisTransaction

Identifier

Intent

A Signing Intent

Jwk

Struct that contains info for a JWK. A list of them for different kids can be retrieved from the JWK endpoint (e.g. https://www.googleapis.com/oauth2/v3/certs). The JWK is used to verify the JWT token.

JwkId

Key to identify a JWK, consists of iss and kid.

MakeMoveVector

MergeCoins

ModifiedAtVersion

MoveCall

The command for calling a Move function, either an entry function or a public function (which cannot return references).

MoveLocation

MultisigAggregatedSignature

The struct that contains signatures and public keys necessary for authenticating a Multisig.

MultisigCommittee

MultisigMember

Object

ObjectDigest

ObjectId

ObjectReference

ObjectReferenceWithOwner

PasskeyAuthenticator

An passkey authenticator with parsed fields. See field defition below. Can be initialized from [struct RawPasskeyAuthenticator].

PasskeyPublicKey

PersonalMessage

ProgrammableTransaction

A series of commands where the results of one command can be used in future commands

Publish

RandomnessStateUpdate

Secp256k1PublicKey

Secp256k1Signature

Secp256r1PublicKey

Secp256r1Signature

SignedCheckpointSummary

SignedTransaction

SplitCoins

StructTag

SystemPackage

Transaction

TransactionDigest

TransactionEffectsDigest

TransactionEffectsV1

The response from processing a transaction or a certified transaction

TransactionEffectsV2

The response from processing a transaction or a certified transaction

TransactionEvents

TransactionEventsDigest

TransferObjects

TypeOrigin

Identifies a struct and the module it was defined in

TypeParseError

UnchangedSharedObject

UnresolvedGasPayment

UnresolvedInputArgument

UnresolvedObjectReference

UnresolvedProgrammableTransaction

UnresolvedTransaction

Upgrade

UpgradeInfo

Upgraded package info for the linkage table

ValidatorAggregatedSignature

ValidatorCommittee

ValidatorCommitteeMember

ValidatorSignature

VersionAssignment

ZkLoginAuthenticator

An zk login authenticator with all the necessary fields.

ZkLoginInputs

All inputs required for the zk login proof verification and other public inputs.

ZkLoginProof

The struct for zk login proof.

ZkLoginPublicIdentifier

A wrapper struct to retrofit in [enum PublicKey] for zkLogin. Useful to construct [struct MultiSigPublicKey].

Enums

Argument

An argument to a programmable transaction command

CheckpointCommitment

Command

A single command in a programmable transaction.

CommandArgumentError

ConsensusDeterminedVersionAssignments

EndOfEpochTransactionKind

EndOfEpochTransactionKind

ExecutionError

ExecutionStatus

IdOperation

InputArgument

IntentAppId

IntentScope

IntentVersion

MultisigMemberPublicKey

MultisigMemberSignature

ObjectData

ObjectIn

If an object exists (at root-level) in the store prior to this transaction, it should be Exist, otherwise it’s NonExist, e.g. wrapped objects should be NonExist.

ObjectOut

ObjectType

Type of a Sui object

Owner

PackageUpgradeError

SignatureScheme

SimpleSignature

TransactionEffects

The response from processing a transaction or a certified transaction

TransactionExpiration

TransactionKind

TypeArgumentError

TypeTag

UnchangedSharedKind

UnresolvedInputArgumentKind

UnresolvedValue

UserSignature

Type Aliases

CheckpointSequenceNumber

CheckpointTimestamp

EpochId

ProtocolVersion

SigningDigest

StakeUnit

Version